Move hexToRgb, rgbToHex and  adjustColor to utils/colorUtils.js
